Reorganize location-related search models
