--- app/models/installation_search.rb
@@ -1,34 +1,0 @@
-class InstallationSearch
-  def initialize(params)
-    @params = params
-  end
-
-  def error_message
-    'There was a problem locating that installation. Mind trying your search again?'
-  end
-
-  def query
-    search_params[:query]
-  end
-
-  def result
-    {
-      latitude: installation.location.latitude,
-      longitude: installation.location.longitude
-    }
-  end
-
-  def valid?
-    search_params.permitted? && installation
-  end
-
-  private
-
-  def installation
-    @installation ||= Installation.find_by('name ILIKE ?', "%#{search_params[:query]}%")
-  end
-
-  def search_params
-    @search_params ||= @params.permit(:query)
-  end
-end
